+++ src/main.cpp
@@ -8,13 +8,14 @@ int main() {
 
 	window.create(800, 600, "Constellation Engine");
 
+	window.clearColor = { 0.5f, 0.0f, 0.4f, 1.0f };
+
 	while (!window.shouldClose) {
+		window.clear();
+		// graphics.render();
 		window.update();
 	}
 
 	window.cleanup();
-
-	// Terminate GLFW, clearing any resources allocated by GLFW
-	glfwTerminate();
 	return 0;
 }